Rename change_set to changeset in integration tests
