+++ integration-tests/deployment/ccip/changeset/1_initial_deploy.go
@@ -1,4 +1,4 @@
-package change_set
+package changeset
 
 import (
 	"github.com/smartcontractkit/chainlink/integration-tests/deployment"
