+++ integration-tests/deployment/ccip/changeset/1_initial_deploy_test.go
@@ -1,4 +1,4 @@
-package change_set
+package changeset
 
 import (
 	"context"
